requirements.txt
requests
feedparser
python-dotenv
APScheduler
